⚡ incremental updates (project structure)

+++ src/code/opcode.hpp
@@ -105,8 +105,8 @@ enum opcode_type: u8 {
 struct opcode {
     u8  op;   // opcode
     u16 fidx; // source code file index
-    u64 num;  // immediate num
-    u64 line; // location line of source code
+    u32 num;  // immediate num
+    u32 line; // location line of source code
     opcode() = default;
     opcode(const opcode&) = default;
     opcode& operator=(const opcode&) = default;
